general changes

+++ ms-credit/src/main/java/com/example/ms_credit/model/CreditEvaluation.java
@@ -20,5 +20,5 @@ public class CreditEvaluation {
     private Boolean applicantAge;
     private Boolean savingCapacity;
 
-    private int creditId;
+    private Long creditId;
 }
